data_process/wholepose/requirements.txt
pytorch
torchvision
opencv
yacs
natsort